refactor std_bounties tests
